Move SemVer definition under $defs (#6)
**Note:** The goal of these changes is to make the schema definition
friendlier for code generation, while keeping the overall schema
unchanged. Concretely, with `$defs`, code generation tools can better
infer the Go struct names.

-- 

This pull request updates the semantic versioning (semver) schema and
its references in the summary schema to improve modularity and
correctness. The main change is the introduction of a `$defs` section
with a `SemVer` definition in the semver schema, and updating dependent
schemas to reference this definition directly.

Schema modularization and reference updates:

* Added a `$defs` section in `schemas/semver/2.0.0/schema.json`
containing a `SemVer` definition for semantic versioning validation, and
set the root `$ref` to this definition.
[[1]](diffhunk://#diff-88a12d9a0cd1921ae6ce6e25c85956d7bb26df434a8bcae735bc1180cd537e45R6-R8)
[[2]](diffhunk://#diff-88a12d9a0cd1921ae6ce6e25c85956d7bb26df434a8bcae735bc1180cd537e45R25-R26)
* Updated `schemas/summary/1.0.0/schema.json` to reference
`#/$defs/SemVer` from the semver schema for both the `version` and
`k6Version` fields, ensuring more precise and modular schema validation.
[[1]](diffhunk://#diff-a0152f74527fc51948306e29b863d15095b70509cbef1eb0d441a6d62a63dbd5L10-R10)
[[2]](diffhunk://#diff-a0152f74527fc51948306e29b863d15095b70509cbef1eb0d441a6d62a63dbd5L25-R25)