Drop `summary.results.passed` (#8)
**Note:** We agreed to park `passed` for now, as we generally exit
early, or throw exceptions, so most of the times it's going to be true
anyway, as k6 needs to finish the execution to in fact produce a
summary. And for all those cases where it could be `false` (checks,
thresholds), the reasoning is complex enough that using a binary signal
wouldn't help much anyway.

-- 

This pull request updates the `schemas/summary/1.0.0/schema.json` file
to simplify the schema for test execution results. The main change is
the removal of the `passed` field from the required properties and from
the schema definition, focusing the results object solely on test
metrics.

Schema simplification:

* Removed the `passed` boolean property from the `results` object, so
test pass/fail status is no longer part of the schema.
* Updated the `required` array for the `results` object to only require
`metrics`, reflecting the removal of the `passed` property.